pyproject.toml
[project]
name = "gbench-v2"
version = "0.1.0"
description = "通用、轻量、易于扩展的大模型评测框架"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
gbench = "gbench.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
